+++ py-polars/polars/datatypes.py
@@ -1,6 +1,11 @@
 import ctypes
 
-__pdoc__ = {"dtype_to_ctype": False, "dtype_to_int": False}
+__pdoc__ = {
+    "dtype_to_ctype": False,
+    "dtype_to_int": False,
+    "dtype_to_primitive": False,
+    "pytype_to_polars_type": False,
+}
 
 
 class DataType: